+++ src/ispd/arquivo/xml/IconicoXML.java
@@ -11,10 +11,10 @@ import ispd.gui.PickModelTypeDialog;
 import ispd.gui.iconico.Edge;
 import ispd.gui.iconico.Vertex;
 import ispd.gui.iconico.grade.VirtualMachine;
-import ispd.motor.carga.CollectionWorkloadGenerator;
-import ispd.motor.carga.TraceFileWorkloadGenerator;
-import ispd.motor.carga.GlobalWorkloadGenerator;
-import ispd.motor.carga.WorkloadGenerator;
+import ispd.motor.carga.workload.CollectionWorkloadGenerator;
+import ispd.motor.carga.workload.TraceFileWorkloadGenerator;
+import ispd.motor.carga.workload.GlobalWorkloadGenerator;
+import ispd.motor.carga.workload.WorkloadGenerator;
 import ispd.motor.filas.RedeDeFilas;
 import ispd.motor.filas.RedeDeFilasCloud;
 import org.w3c.dom.Document;
