+++ src/ispd/arquivo/xml/models/builders/LoadBuilder.java
@@ -2,12 +2,12 @@ package ispd.arquivo.xml.models.builders;
 
 import ispd.arquivo.xml.utils.WrappedDocument;
 import ispd.arquivo.xml.utils.WrappedElement;
-import ispd.motor.carga.CollectionWorkloadGenerator;
-import ispd.motor.carga.PerNodeWorkloadGenerator;
-import ispd.motor.carga.GlobalWorkloadGenerator;
-import ispd.motor.carga.TraceFileWorkloadGenerator;
-import ispd.motor.carga.WorkloadGenerator;
-import ispd.motor.carga.WorkloadGeneratorType;
+import ispd.motor.carga.workload.CollectionWorkloadGenerator;
+import ispd.motor.carga.workload.PerNodeWorkloadGenerator;
+import ispd.motor.carga.workload.GlobalWorkloadGenerator;
+import ispd.motor.carga.workload.TraceFileWorkloadGenerator;
+import ispd.motor.carga.workload.WorkloadGenerator;
+import ispd.motor.carga.workload.WorkloadGeneratorType;
 import ispd.motor.carga.task.TaskSize;
 
 import java.io.File;
