+++ src/ispd/motor/carga/task/RandomicTaskBuilder.java
@@ -1,6 +1,5 @@
-package ispd.motor.carga;
+package ispd.motor.carga.task;
 
-import ispd.motor.carga.task.TaskSize;
 import ispd.motor.filas.Tarefa;
 import ispd.motor.filas.servidores.CS_Processamento;
 import ispd.motor.random.Distribution;
