+++ src/ispd/motor/carga/workload/CollectionWorkloadGenerator.java
@@ -1,4 +1,4 @@
-package ispd.motor.carga;
+package ispd.motor.carga.workload;
 
 import ispd.motor.filas.RedeDeFilas;
 import ispd.motor.filas.Tarefa;
