+++ src/ispd/motor/carga/workload/WorkloadGeneratorType.java
@@ -1,4 +1,4 @@
-package ispd.motor.carga;
+package ispd.motor.carga.workload;
 
 /**
  * Enumeration of all possible types of {@link WorkloadGenerator}s.
